requirements.txt
aiohttp==3.9.1
selenium==4.16.0
beautifulsoup4==4.12.2
webdriver_manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
snscrape==0.7.0.20230622
pandas==2.1.4
textblob==0.17.1
